- Broke parts of runClient() into two other functions, making it much clearer   and easier to read. - Added Inited<> and Zeroed<> template classes that let you specify how   a class member should be initialized.  See Utility.h for documentation. - Made the joystick axis values normalized floats in the range [-1, 1] for   the sake of clarity and consistency with other Input objects.
This change should reduce the total number of lines of code.  Almost
always a good thing.  :)


git-svn-id: file:///desktop/SourceForge%20Backup/empyrean/svnrepos/trunk@322 fea1f14b-4e95-431d-9825-d582385f3f73

+++ empyrean/src/client/InputManager.h
@@ -26,7 +26,7 @@ namespace pyr {
         void onMousePress(Uint8 button, bool down, int x, int y);
         void onMouseMove(int x, int y);
         void onJoyPress(Uint8 button, bool down);
-        void onJoyMove(int axis, int value);
+        void onJoyMove(int axis, float value);
 
     private:
         static std::string getSDLKeyName(SDLKey key);
